move store models to correct place
