+++ src/app/store/models.ts
@@ -1,1 +1,3 @@
+export * from './models/delete-status.model';
+export * from './models/deleted-node-info.model';
 export * from './models/node-info.model';
