+++ src/app/store/models/deleted-node-info.model.ts
